requirements.txt
sqlalchemy==2.0.27
pandas==2.1.3
python-dotenv==1.0.0
